Former-commit-id: f1c1453a0524101f6903d8b0248c9b3d191f04fe

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/QuestionDisabled.cs
@@ -5,7 +5,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 {
     public class QuestionDisabled : QuestionPassiveEvent
     {
-        public QuestionDisabled(Guid questionId, int[] propagationVector)
+        public QuestionDisabled(Guid questionId, decimal[] propagationVector)
             : base(questionId, propagationVector) {}
     }
 }
